\name{dimensoes}
\alias{dimensoes}
\title{Federal Budget Dimensions}
\description{
  Here, we list the dimensions concerning Federal Budget expenditure. For additional information, please refer to the 'Manual Técnico do Orçamento'.

  PT-BR: Aqui, listamos as dimensões relativas aos gastos do Orçamento Federal. Para informações adicionais, consulte o 'Manual Técnico do Orçamento'.
}
\section{Classificação do Orçamento Federal - dimensões}{
\tabular{lll}{
Parameter\tab  \strong{Classification by Budget segment} \tab \strong{Classificação por Esfera}\cr
\code{Esfera} \tab Budget segment \tab Esfera Orçamentária \cr
.\cr
\tab \strong{Institutional Classification} \tab \strong{Classificação Institucional}\cr
\code{Orgao} \tab Agency \tab Órgão \cr
\code{UO} \tab Budgetary Unit \tab Unidade Orçamentária \cr
.\cr
\tab \strong{Functional Classification} \tab \strong{Classificação Funcional}\cr
\code{Funcao} \tab Function \tab Função \cr
\code{Subfuncao} \tab Subfunction \tab Subfunção \cr
. \cr
\tab \strong{Programmatic Structure} \tab \strong{Classificação Programática}\cr
\code{Programa} \tab Program \tab Programa \cr
\code{Acao} \tab Activity \tab Ação \cr
\code{PlanoOrcamentario} \tab Budget Plan \tab Plano orçamentário \cr
\code{Subtitulo} \tab Subtitle \tab Subtítulo \cr
.\cr
\tab \strong{Detailed Financial Programming} \tab \strong{Componentes da programação financeira}\cr
\code{CategoriaEconomica} \tab Economic Category \tab Categoria Econômica \cr
\code{GND} \tab Nature of Expenditure \tab Grupo Natureza da Despesa \cr
\code{ModalidadeAplicacao} \tab Application Mode \tab Modalidade de Aplicação \cr
\code{ElementoDespesa} \tab Expenditure Element \tab Elemento da Despesa \cr
.\cr
\code{Fonte} \tab Source of Funds \tab Fonte de Recursos \cr
\code{IdUso} \tab Resources used as a counterpart Identifier \tab Identificador de recursos utilizado como contrapartida \cr
\code{ResultadoPrimario} \tab Primary Result Identifier \tab Identificador de efeito sobre o resultado primário da União \cr
}
}
